+++ changes.md
@@ -3,6 +3,7 @@
 - **API breaking change**: update signature of `onyx.api/await-job-completion` to take an opts map.
 - Allow functions in leaf position of a workflow. [#198](https://github.com/onyx-platform/onyx/issues/198)
 - Bug fix: flow-conditions retry default action should emit segments [#262](https://github.com/onyx-platform/onyx/issues/262)
+- API: New catalog entry option `:onyx/n-peers` to automatically expand to make `:onyx/min-peers` and `:onyx/max-peers` peers the same value. [#282](https://github.com/onyx-platform/onyx/issues/282)
 
 #### 0.7.3